Adds basic cmake support
You may build the QUaServer library and the examples using cmake.
As for now, the dependencies have to be build separately and have to be
installed onto the build system.

The cmake build is currently only tested on an recent Ubuntu system!

Requirements:
- open62541
- libmbedtls-dev (needed to build open62541)

Later on the implicit build of the dependecies may be supported.